+++ changelog.md
@@ -7,6 +7,7 @@
   parameters. The `workingDirectory` parameter is now deprecated in all methods, 
   use `RunOptions.workingDirectory` instead.
 - Added `PubApp.runAsync`.
+- In `Dart.run`, deprecate the vmNewGenHeapMB and vmOldGenHeapMB options.
 
 ## 0.7.0
 - Big changes! Task definititions are now primarily annotation based; see the